// src/types/marketplace-api.ts
export interface components {
    schemas: {
// ...
        AddonCollection: {
            _links: components["schemas"]["AddonCollectionLinks"];
            _embedded: components["schemas"]["AddonCollectionEmbedded"];
            /**
             * Format: int32
             * @description The total number of apps matched by the query
             */
            count: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}